PIG-1135: skewed join partitioner returns negative partition index  (yinghe via olgan)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/pig/trunk@889305 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/backend/hadoop/executionengine/physicalLayer/relationalOperators/POPartitionRearrange.java
@@ -223,7 +223,7 @@ public class POPartitionRearrange extends POLocalRearrange {
 			Tuple opTuple = mTupleFactory.newTuple(4);
 			opTuple.set(0, t.get(0));
 			// set the partition index
-			opTuple.set(1, reducerIdx.byteValue());
+			opTuple.set(1, reducerIdx.intValue());
 			opTuple.set(2, key);
 			opTuple.set(3, t.get(2));
